Remove duplicate listener🤦‍♂
